Create README.md
The Weather Forecast & Rain Prediction App is a web application powered by machine learning that predicts temperature, humidity, and the likelihood of rain. The app provides users with both real-time and future weather information. The app leverages machine learning models to forecast temperature and humidity and uses a classification model to predict rain